requirements.txt.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt


Flask==1.0.2
mysql-connector-python==8.0.19
requests==2.23.0
Flask-Cors==3.0.8